Implementing case-insensitive columns on import
Implemented case-insensitive flag for database models. This not only simplifies when for instance searching for information, but also decreases query times when performing searches on the index. The actual case is preseved in the fields but when doing searches the query is interpreted as case-insensitive. This have only been added where it seems logical to do so in the models but can be changed by toggeling the nocase attribute in the models.